requirements.txt
Flask==2.1.0
flask-restx==1.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
marshmallow==3.16.0
PyJWT==2.4.0
pytest==7.4.4
SQLAlchemy==1.3.24
Werkzeug==2.1.2
jinja2==3.1.3
markupsafe==2.1.3
itsdangerous==2.1.2